Move version functions from utils to __version__ module
Also stop using AugurException in node file parsing. It doesn't add any
information over RuntimeError.
