+++ augur/util_support/node_data_file.py
@@ -1,8 +1,7 @@
 import json
 
-from augur.utils import AugurException  # TODO use something more meaningful
-from augur.utils import get_augur_version
-from augur.utils import is_augur_version_compatible
+from augur.__version__ import __version__
+from augur.__version__ import is_augur_version_compatible
 from augur.validate import validate_json, ValidateError, load_json_schema
 
 
@@ -59,21 +58,21 @@ class NodeDataFile:
                     self.fname,
                 )
             except ValidateError as err:
-                raise AugurException(
+                raise RuntimeError(
                     f"{self.fname} contains an `annotations` attribute of an invalid JSON format. Was it "
                     "produced by different version of augur the one you are currently using "
-                    f" ({get_augur_version()})? Please check the program that produced that JSON file."
+                    f" ({__version__})? Please check the program that produced that JSON file."
                 ) from err
 
         if not isinstance(self.nodes, dict):
-            raise AugurException(
+            raise RuntimeError(
                 f"`nodes` value in {self.fname} is not a dictionary. Please check the formatting of this JSON!"
             )
 
         if self.is_generated_by_incompatible_augur:
-            raise AugurException(
+            raise RuntimeError(
                 f"Augur version incompatibility detected: the JSON {self.fname} was generated by "
                 f"{self.generated_by}, which is incompatible with the current augur version "
-                f"({get_augur_version()}). We suggest you rerun the pipeline using the current version of "
+                f"({__version__}). We suggest you rerun the pipeline using the current version of "
                 "augur."
             )
